Add scripts for comparing defconfigs across branches
When porting a defconfig from <branch a> to <branch b>, it can be helpful
to easily compare a defconfig (e.g. imx8_var_defconfig) and the defconfig
it was derived from (e.g. imx_v8_defconfig) on an old branch to know
how to port it on a new branch.

These scripts help to automate the development and review of porting defconfigs.

Signed-off-by: Nate Drude <nate.d@variscite.com>